fix memory leak in byte-class creation
Primary issue is the creation of the byte-class that was stored in an
expr was not created in the arena.

Additionally, have added functions to RangeSet to allow deinit to be
called and clean-up resources. This isn't used internally as-is due to
relying on the arena.

Expression creation is moved into a dedicated function to nudge towards
correctly using the arena to create expression objects.